Add menu drawer
I added a new button menu that users can click on whenever they are on
mobile phones (i used <768 as a breakpoint).
When the button is clicked a drawer will be shown with the navbar items
so that the user can navigate around the page.
I also created a reusable shadow component that can be used with
modals or drawers.